Удалил временные файлы, исправил ошибку линтера
